requirements.txt
alembic==1.12.0
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
flask==2.3.3
Flask-Markdown==0.3
Flask-Migrate==4.0.4
flask-sqlalchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing-extensions==4.7.1
werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2